refactor: lógica de excluir e marcar como lida as pergunstas  :white_check_mark:

+++ components/Section/Questions.vue
@@ -1,23 +1,28 @@
 <template>
   <section>
-    <div v-if="loading">
-      <p>Carregando perguntas...</p>
+    <div v-if="loading" class="fixed inset-0 bg-slate-900 bg-opacity-80 flex justify-center items-center z-50">
+      <Spinner/>
     </div>
     <div v-else-if="questions.length > 0">
-      <div v-for="question in questions" :key="question.id" class="bg-light-blue my-5 p-5">
-        <div class="flex items-center gap-3">
-          <div class="user bg-blue-100 w-10 p-2 mx-0 rounded-full">
-            <User color="#ffff" />
+      <div v-for="question in questions" :key="question.id"
+        class="my-5 p-5 bg-light-blue relative shadow-sm rounded-md">
+        <div :class="{ 'blur-sm ': question.isRead }">
+          <div class="flex items-center gap-3">
+            <div class="user bg-blue-100 w-10 p-2 mx-0 rounded-full">
+              <User color="#ffff" />
+            </div>
+            <p class="text-blue">{{ question.content }}</p>
           </div>
-          <p class="text-blue">{{ question.content }}</p>
+          <div class="flex justify-end items-center gap-6 mt-3 px-24">
+            <button v-if="!question.isRead" @click="openCheckModal(question)" class="flex items-center">
+              <CheckIcon :size="17" color="#3485ff" />
+              <p class="text-blue-100">Marcar como lida</p>
+            </button>
+          </div>
         </div>
-        <div class="flex justify-end items-center gap-6">
-          <button @click="showModalCheck = true" class="flex items-center">
-            <CheckIcon :size="18" color="#3485ff" />
-            <p class="text-blue-100">Marcar como lida</p>
-          </button>
+        <div class="absolute bottom-5 right-5">
           <button @click="openDeleteModal(question)" class="flex items-center">
-            <Trash :size="18" color="#E83F5B" />
+            <Trash :size="17" color="#E83F5B" />
             <p class="text-[#E83F5B]">Excluir</p>
           </button>
         </div>
@@ -28,7 +33,8 @@
       <p>Nenhuma pergunta por aqui...</p>
       <p>Faça sua primeira pergunta e envie o <br>código dessa sala para outras pessoas!</p>
     </div>
-    <Check :isVisible="showModalCheck" @close="showModalCheck = false" />
+    <Check :isVisible="showModalCheck" :question="questionToMarkAsRead" @close="showModalCheck = false"
+      @markAsRead="handleMarkAsRead" />
     <Delete :isVisible="showModalDelete" :question="questionToDelete" @close="showModalDelete = false"
       @deleteQuestion="handleDeleteQuestion" />
   </section>
@@ -38,6 +44,7 @@
 p {
   @apply text-blue;
 }
+
 </style>
 
 <script setup>
@@ -46,10 +53,14 @@ import { User, Trash, CheckIcon } from 'lucide-vue-next'
 import { useQuestion } from '@/composables/useQuestion'
 import Check from '@/components/Modal/Check.vue';
 import Delete from '@/components/Modal/Delete.vue';
+import { useToast } from 'vue-toastification'
+import Spinner from '@/components/Loading/Spinner.vue';
 
+const toast = useToast();
 const showModalCheck = ref(false);
 const showModalDelete = ref(false);
-const questionToDelete = ref(null); // Armazena a pergunta a ser excluída
+const questionToDelete = ref(null);
+const questionToMarkAsRead = ref(null);
 const route = useRoute()
 const roomCode = ref(route.params.code)
 const questions = ref([])
@@ -64,13 +75,54 @@ const fetchQuestions = async () => {
   } catch (error) {
     console.error('Erro ao buscar perguntas:', error)
   } finally {
-    loading.value = false
+    setTimeout(() => {
+      loading.value = false
+    }, 300) 
   }
 }
 
-const markAsRead = async (questionId) => {
-  await updateQuestion(questionId, true)
-  await fetchQuestions()
+const openCheckModal = (question) => {
+  questionToMarkAsRead.value = question;
+  showModalCheck.value = true;
+}
+
+const handleMarkAsRead = async ({ question, password }) => {
+  if (!question || !question.id) {
+    console.error("Pergunta não definida ou inválida:", question);
+    return;
+  }
+
+  try {
+    const result = await updateQuestion({
+      id: question.id,
+      isRead: true,
+      roomCode: roomCode.value,
+      password,
+    });
+
+    if (result.success) {
+      const questionToUpdate = questions.value.find(q => q.id === question.id);
+      if (questionToUpdate) {
+        questionToUpdate.isRead = true;
+      }
+      toast.success('Pergunta marcada como lida com sucesso!', {
+        timeout: 3100,
+        position: 'bottom-right',
+        pauseOnFocusLoss: false,
+        pauseOnHover: false,
+        closeButton: false,
+      })
+    }
+
+  } catch (error) {
+    toast.error('Erro ao marcar pergunta como lida. Tente novamente mais tarde.', {
+        timeout: 3100,
+        position: 'bottom-right',
+        pauseOnFocusLoss: false,
+        pauseOnHover: false,
+        closeButton: false,
+      })
+  }
 }
 
 const openDeleteModal = (question) => {
@@ -79,20 +131,38 @@ const openDeleteModal = (question) => {
 }
 
 const handleDeleteQuestion = async ({ question, password }) => {
+  if (!question || !question.id) {
+    console.error("Pergunta não definida ou inválida:", question);
+    return;
+  }
+
   try {
     const result = await deleteQuestion({
       id: question.id,
       roomCode: roomCode.value,
-      password, // Envia a senha para o backend
+      password,
     });
 
     if (result.success) {
       questions.value = questions.value.filter(q => q.id !== question.id);
+      toast.success('Pergunta excluída com sucesso!', {
+        timeout: 3100,
+        position: 'bottom-right',
+        pauseOnFocusLoss: false,
+        pauseOnHover: false,
+        closeButton: false,
+      })
     } else {
-      console.error(`Falha ao excluir pergunta:`, result.error);
+      throw new Error(result.error || 'Falha ao excluir pergunta');
     }
   } catch (error) {
-    console.error(`Erro ao excluir pergunta:`, error);
+    toast.error(error.message || 'Erro ao excluir pergunta. Verifique a senha e tente novamente.', {
+      timeout: 3100,
+      position: 'bottom-right',
+      pauseOnFocusLoss: false,
+      pauseOnHover: false,
+      closeButton: false,
+    })
   }
 }
